+++ src/editor/core/plugins/emoji/provider.js
@@ -137,7 +137,10 @@ class EmojiChooser {
                     break;
             }
         }).on('keyup', function(e) {
-            if (e.which !== 8 && !/[a-z0-9\d]/i.test(String.fromCharCode(e.which))) {
+            let keyCode = e.keyCode || e.which;
+            // This line should prevent processing in case user presses down/up on desktop, android chrome does not send
+            // always send keyCode 229 so we can skip this check in this case
+            if (keyCode !== 229 && keyCode !== 8 && !/[a-z0-9\d]/i.test(String.fromCharCode(keyCode))) {
                 return;
             }